#register 1. register -> signup 이름 수정. 2. 이미 회원인지 확인하는 상태관리 추가.
